requirements.txt
torch
torchvision
numpy
pillow
matplotlib
streamlit
streamlit-drawable-canvas
